requirements.txt
flask
flask-cors
pillow
google-generativeai
os